include: flash_map: Add multi image definitions to flash map
Adds definitions for `image_2` and `image_3` to the flash map. When we
are using the multi-image build which has `mcuboot_primary_1` and
`mcuboot_secondary_1`.

There is special handling for `image_2` for `nRF53` since it only
exsists in RAM and only when flash simulator is turned on will it be
available for the applications using it.

The application image does not need to be aware of `image_2` for nRF53
as this is a emulated flash of the network core and would only consume
RAM from the application. The application only needs to be aware of
`image_3` which is where it can store the network core image for MCUBoot
to upgrade on the next reboot.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>