applications: Created application bridged devices creator module.
The bridged devices adding and removing was handled in the shell
module, however it may be useful to trigger such actions from
a different parts of an application (e.g. app task).

Extracted methods to add and remove bridged devices and handle
Bluetooth LE connected callback to the dedicated bridged devices
creator module. It will be possible to use its API in shell and
in the different locations as well.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>